[cmake] Fix dependency for Interpreter.
