+++ include/glow/Quantization/Quantization.h
@@ -50,28 +50,6 @@ struct NodeQuantizationInfo {
   }
 };
 
-/// A data structure that represents the 32-bit to 8-bit quantization
-/// scaling operation. This data structure represents the transformation:
-/// (((input >> pre) * scale) + rtn) >> post + offset.
-struct QuantizationTransform32To8 {
-  int pre_;
-  int post_;
-  int scale_;
-  int offset_;
-
-  /// Initializes the transformation based on the conversion formula (above).
-  QuantizationTransform32To8(int pre, int post, int scale, int offset)
-      : pre_(pre), post_(post), scale_(scale), offset_(offset) {}
-
-  /// \returns the scaled integer.
-  int32_t transform(int32_t input) {
-    // The operation x >> y is rounded down to negative infinity. To get to
-    // round-nearest we add (1 << (shift - 1)) to the value prior to shifting.
-    int rtn = (1 << (post_ - 1));
-    return ((((input >> pre_) * scale_) + rtn) >> post_) + offset_;
-  }
-};
-
 namespace quantization {
 
 /// Calculate TensorQuantizationParams based on the clipped \p min and \p max
@@ -81,14 +59,6 @@ TensorQuantizationParams chooseQuantizationParams(float min, float max);
 /// Generate NodeQuantizationInfo for all required nodes from graph \p G.
 std::vector<NodeQuantizationInfo>
 generateNodeQuantizationInfos(const Function *F);
-
-/// Convert the floating point quantization parameters \p scale and \p offset
-/// into the integer sequence of:
-/// result = ((input >> pre) * scale) >> post + offset.
-/// This scales a 32-bit signed integer word into an 8-bit signed integer.
-/// \returns transformation parameters.
-QuantizationTransform32To8 quantizeScaleOffset32To8(float scale,
-                                                    int32_t offset);
 
 /// Quantizes the function \p F into a new unoptimized partially quantized
 /// function based on \p quantizationInfos. This method converts to integer as
@@ -101,7 +71,6 @@ quantizeFunction(const ExecutionEngine &EE,
                  Function *F, llvm::StringRef newFuncName = "");
 
 } // namespace quantization
-
 } // namespace glow
 
-#endif
+#endif // GLOW_QUANTIZATION_QUANTIZATION_H
