+++ lib/Backends/Interpreter/CMakeLists.txt
@@ -7,4 +7,4 @@ target_link_libraries(Interpreter
                         Base
                         Graph
                         IR
-                        Quantization)
+                        QuantizationBase)
