+++ lib/Optimizer/CMakeLists.txt
@@ -9,4 +9,5 @@ target_link_libraries(Optimizer
                       PRIVATE
                         Graph
                         IR
-                        Backends)
+                        Backends
+                        QuantizationBase)
