+++ lib/Quantization/Base/CMakeLists.txt
@@ -1,6 +1,4 @@
-add_library(QuantizationBase
-		QuantizationBase.cpp)
 
-target_link_libraries(QuantizationBase
-                      PUBLIC
-                        Support)
+add_library(QuantizationBase
+              Profile.cpp
+              Base.cpp)
