+++ lib/Quantization/CMakeLists.txt
@@ -1,5 +1,7 @@
+
+add_subdirectory(Base)
+
 add_library(Quantization
-              Profile.cpp
               Serialization.cpp
               Quantization.cpp)
 
@@ -7,4 +9,4 @@ target_link_libraries(Quantization
                       PRIVATE
                         Graph
                         ExecutionEngine
-			QuantizationBase)
+                        QuantizationBase)
